pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nonebot-plugin-distributed-blacklist"
version = "1.0.0"
description = "基于PostgreSQL的分布式黑名单系统"
authors = [
    {name = "Tosd0", email = "tntobsidian@126.com"}
]
dependencies = [
    "nonebot2>=2.3.0,<3.0.0",
    "nonebot-adapter-onebot>=2.3.0",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-localstore>=0.6.0",
    "asyncpg>=0.28.0",
    "pydantic>=1.10.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "LGPL-3.0"}
keywords = ["nonebot", "blacklist", "distributed", "postgresql"]
classifiers = [
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/Tosd0/nonebot-plugin-distributed-blacklist"
Repository = "https://github.com/Tosd0/nonebot-plugin-distributed-blacklist"
Documentation = "https://github.com/Tosd0/nonebot-plugin-distributed-blacklist"
Issues = "https://github.com/Tosd0/nonebot-plugin-distributed-blacklist/issues"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio", 
    "black",
    "isort",
    "flake8",
]

[tool.setuptools.packages.find]
include = ["nonebot_plugin_distributed_blacklist*"]

